refactor(handler) split codec into injector and extractor. inline injector
