+++ kong-plugin-zipkin-0.2.1-1.rockspec
@@ -22,7 +22,7 @@ dependencies = {
 build = {
   type = "builtin";
   modules = {
-    ["kong.plugins.zipkin.codec"] = "kong/plugins/zipkin/codec.lua";
+    ["kong.plugins.zipkin.extractor"] = "kong/plugins/zipkin/extractor.lua";
     ["kong.plugins.zipkin.handler"] = "kong/plugins/zipkin/handler.lua";
     ["kong.plugins.zipkin.random_sampler"] = "kong/plugins/zipkin/random_sampler.lua";
     ["kong.plugins.zipkin.reporter"] = "kong/plugins/zipkin/reporter.lua";
